Sort out test environment setup
