+++ e2e/node/acp-v3.test.ts
@@ -20,17 +20,15 @@
  */
 
 import {
-  jest,
-  beforeAll,
   beforeEach,
   afterEach,
   describe,
   it,
+  test,
   expect,
 } from "@jest/globals";
 
 import { Session } from "@inrupt/solid-client-authn-node";
-import { config } from "dotenv-flow";
 import {
   getSolidDataset,
   setThing,
@@ -38,47 +36,38 @@ import {
   overwriteFile,
   getFile,
   getSourceUrl,
-  deleteFile,
   createSolidDataset,
   deleteSolidDataset,
   UrlString,
   acp_v3 as acp,
   FetchError,
 } from "../../src/index";
-import { getTestingEnvironment, TestingEnvironment } from "./util/getTestingEnvironment";
-import { getAuthenticatedSession } from "./util/getAuthenticatedSession";
+import { getTestingEnvironment, TestingEnvironment } from "../util/getTestingEnvironment";
+import { getAuthenticatedSession } from "../util/getAuthenticatedSession";
 
-let env: TestingEnvironment;
-let options: { fetch: typeof global.fetch };
-let session: Session;
-let sessionResource: string;
+const env: TestingEnvironment = getTestingEnvironment();
 const sessionResourcePrefix: string = "solid-client-tests/node/acp-v3-";
+if (env.feature.acp_v3 !== true) {
+  test.only(`Skipping unsupported ACP V3 tests in ${env.environment}`, () => {});
+}
 
-beforeAll(() => {
-  config({
-    path: __dirname,
-    // Disable warning messages in CI
-    silent: process.env.CI === "true",
+describe("Authenticated end-to-end ACP V3", () => {
+  let options: { fetch: typeof global.fetch };
+  let session: Session;
+  let sessionResource: string;
+  
+  beforeEach(async () => {
+    session = await getAuthenticatedSession(env);
+    sessionResource = `${env.pod}${sessionResourcePrefix}${session.info.sessionId}`;
+    options = { fetch: session.fetch };
+    await saveSolidDatasetAt(sessionResource, createSolidDataset(), options);
   });
-  env = getTestingEnvironment();
-  if (env.feature.acp_v3 !== true) {
-    return;
-  }
-});
+  
+  afterEach(async () => {
+    await deleteSolidDataset(sessionResource, options);
+    await session.logout();
+  });
 
-beforeEach(async () => {
-  session = await getAuthenticatedSession(env);
-  sessionResource = `${env.pod}${sessionResourcePrefix}${session.info.sessionId}`;
-  options = { fetch: session.fetch };
-  await saveSolidDatasetAt(sessionResource, createSolidDataset(), options);
-});
-
-afterEach(async () => {
-  await deleteSolidDataset(sessionResource, options);
-  await session.logout();
-});
-
-describe(`Authenticated end-to-end ACP V3`, () => {
   async function initialisePolicyResource(
     policyResourceUrl: UrlString,
     session: Session
