+++ e2e/util/getAuthenticatedSession.ts
