+++ e2e/util/setupEnv.ts
@@ -1,0 +1,7 @@
+import { config } from "dotenv-flow";
+
+config({
+  path: __dirname.concat("/../env/"),
+  // Disable warning messages in CI
+  silent: process.env.CI === "true",
+});
